Multi furniture select query
Added Layout dropdown to refine search for a multi query and prevent
contamination of survey's with different layouts for the sam floor.
Began work on report-multi-survey.php and query_survey_range.js which is
the core files for getting information from the database with which to
populate the map with.