Refined Pulled Data
Refined the Data pulled from the database and performed some basic math
to calculate the average use ratio of each individual furniture across
surveys, the average occupancy of a piece of furniture for that survey,
and the total number of occupants for that piece of furniture across all
surveys. This inofrmation gets stored in a map. TODO: sum furniture
ratios for each area, iterating through both the area map and the
furniture map to calculate the use of a total area.