Line Placement Fix and Modified Data Calculations
Lines are now drawn from the center of an object rather than the top
corner of the object, resizing on zoom is properly maintained and
controlled. Data of Modified furniture not added to primary furniture
object and is instead calculated by where the modified furniture is
placed in the area, adding its values to the proper places and
representing accurate area data.