Improved Print Out Data
Altered get-area-peaks.php to properly obtain area information as they
relate to rooms, and adjusted query_survey_range.js to properly output
the string for a room containing its peak information. Also added
modified furniture to the count of furniture within an area over a
period to properly find ratios of use as it pertains to furniture.